Created the 4 activities of the app: MainActivity, SecondActivity, StatsActivity, SettingsActivity.
MainActivity acts as the main menu for the app, printing the logo and
having 3 buttons that lead to the other activities.

SecondActivity is the activity where beverages are added by the user.
The screen is split in 6 panels (3 rows, 2 columns): 5 panels with
editable drink presets and a panel for custom drinks. At the bottom of
the screen there are two buttons for faster access to StatsActivity
and SettingsActivity.

StatsActivity shows the blood alcohol content (BAC) of the user,
calculated based on body weight entered in SettingsActivity and drinks
added in SecondActivity. A list of the drinks entered is shown,
displaying the image, name, volume and alcohol percentage of every
drink. The effects of alcohol that the user may experience, according to
https://en.wikipedia.org/wiki/Blood_alcohol_content, are also listed.
BAC color changes with its value.

SettingsActivity has a "RESET DRINK PREFERENCES" button that resets
drink autofill values from SecondActivity to default and an EditText
for entering body weight.